requirements.txt
django
djangorestframework
python-dotenv
gunicorn